new post seems to work but not refreshing, cannot update either

+++ public/javascript/delete-post.js
@@ -1,9 +1,21 @@
 async function deleteFormHandler(event) {
-    event.preventDefault();
+
     
-    await fetch(`/api/posts/${id}`, {
-        method: 'DELETE'
-    });
-}
+    if (event.target.classList.contains("delete-post-btn")) {
+        const id = event.target.id
+        console.log(id);
+        const delPost = await fetch(`/api/posts/${id}`, {
+            
+            method: 'DELETE'
+            
+        });
+        if (delPost.ok) {
+            console.log("Deleted successful!");
+            window.location.reload();
+        } else {
+            console.log(delPost.statusText);
+        }
+    }
+}      
   
-document.querySelector('.delete-post-btn').addEventListener('click', deleteFormHandler);
+document.querySelector('.post-section').addEventListener('click', deleteFormHandler);
